+++ packages/orq-rc/RELEASES.md
@@ -8679,3 +8679,13 @@ Based on:
 - [typescript v4.9.0-rc.5] packages/orq-rc
 ### Releases
 - [NPM v4.9.0-rc.5] https://www.npmjs.com/package/@orq-ai/node/v/4.9.0-rc.5 - packages/orq-rc
+
+## 2026-04-28 10:08:17
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.9 (2.881.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.9.0-rc.6] packages/orq-rc
+### Releases
+- [NPM v4.9.0-rc.6] https://www.npmjs.com/package/@orq-ai/node/v/4.9.0-rc.6 - packages/orq-rc